Fix CVXOPT inequality presolve: opposite-row deletion and dual recovery (#3390)
* Only merge positive-ratio rows in CVXOPT inequality presolve

compress_matrix, used by the CVXOPT interface to drop redundant
inequality rows when the equality matrix is rank deficient, treated a
row as redundant whenever it was any scalar multiple of a kept row with
a matching right-hand side ratio. For inequalities Gx <= h a negative
multiple describes the opposite half-space (e.g. z >= 1 stored as
-z <= -1 versus z <= 1), so dropping it enlarges the feasible set:
min z s.t. z <= 1, z >= 1 plus duplicated equalities was reported
unbounded instead of optimal at 1.

Require a positive ratio before merging rows. compress_matrix's only
caller compresses the LEQ block of G in cvxopt_conif.py, so no equality
usage relies on merging negative multiples.

Tests: unit test that compress_matrix keeps negative-multiple rows, and
a CVXOPT regression test for the formerly-unbounded problem.


* Fix dual recovery for inequalities merged by CVXOPT presolve

When CVXOPT's presolve merged duplicate inequality rows, the duals of
the original constraints were recovered as P_leq @ z, giving every
merged row a scaled copy of the full compressed multiplier. Duplicated
constraints then received duals summing to (1 + ratio^2) times the
correct value, violating KKT stationarity (e.g. z >= 1 stated twice
produced duals 1.0 and 1.0 where the sum must be 1.0).

Recover the duals with an expansion matrix that assigns each compressed
dual entirely to the row that was kept and zero to dropped duplicates,
so G_orig.T @ y_orig == G_compr.T @ z and stationarity holds.
compress_matrix now also returns the kept row indices to build this
expansion.

Tests: CVXOPT regression test checking stationarity of the recovered
duals for exact and scaled duplicate inequalities; existing
compress_matrix unit test extended to cover the kept-row indices.


* move CVXOPT-only utility into cvxopt_conif.py

move tests of CVXOPT utility from free-function in test_conic_solvers.py to the class test_conic_solvers.py::TestCVXOPT

* Fixes precommit

---------
